feat: move metadata interface into metadata package
