+++ dfs/internal/coordinator/metadata_mock.go
@@ -2,7 +2,7 @@ package coordinator
 
 import (
 	"github.com/mochivi/distributed-file-system/internal/common"
-	"github.com/mochivi/distributed-file-system/internal/storage"
+	"github.com/mochivi/distributed-file-system/internal/storage/metadata"
 	"github.com/stretchr/testify/mock"
 )
 
@@ -14,7 +14,7 @@ func (m *MockMetadataSessionManager) trackUpload(sessionID string, req common.Up
 	m.Called(sessionID, req, numChunks)
 }
 
-func (m *MockMetadataSessionManager) commit(sessionID string, chunkInfos []common.ChunkInfo, metaStore storage.MetadataStore) error {
+func (m *MockMetadataSessionManager) commit(sessionID string, chunkInfos []common.ChunkInfo, metaStore metadata.MetadataStore) error {
 	args := m.Called(sessionID, chunkInfos, metaStore)
 	return args.Error(0)
 }
