+++ dfs/internal/storage/metadata/types.go
@@ -1,4 +1,4 @@
-package storage
+package metadata
 
 import (
 	"time"
